Classes! major changes to contexts (see full comment)
long note to self follows:

when i first started xxl i imagined a context to be a stack of
scopes (dicts) with a bit of code at the end

this ended up being quite complex and confusing

i'm thinking now that i prefer scopes to work like globals,
wherein the list of defined names is flat, even if you're in a
deeply nested applyexpr call.

under this new understanding, i believe contexts are
better represented as [scopedict,code].

this commit changes and dramatically simplifies the process
of creating a context, doing get/set inside it, etc.

In theory this should be much faster as well, though with
Ackermann I am getting disturbingly slow results these days.

That said, "doc/lang.xxl" load and Workbench are much faster.

to go along with this, i've dropped the context of top-level
selectors like .file.get, for now.

i've also changed the way "getmodular" works, for the better:

in the past, doing . get ['file,"def.h"] would return the
equivalent of .file.get.

this required a lot of extra code, was majorly confusing when
trying to implement dict.key, and was slow.

i've wanted to use the concept of a tag that starts with a capital
letter as class, for some time. this is the first whack at that.

first, know that you can assign your own tag to a value with make,
i.e., [1,2,3]$'mything. You can also use capital letters. Tags
with capital letters perform exactly the same as regular tags,
with one major difference..

the idea is that get and set and a few other methods (probably
"apply") will, when confronted with a value that has a tag assigned,
whose tag is a capital letter (i.e., what we call a Class), look
for Class.[verb] and call that instead of the usual get/set/apply
behavior.

i.e, . get ("def.h"$'File)

so far, only the get part is implemented. i'm still at a loss as to
how to make this fast, but at least it's flexible and I think
pretty logical.

some tests had to be removed because they exercised the old
['file,"def.h"]-style modular gets.