major fixes to make set-uplevel work as expected; w/tests! read this for summary of how contexts work at this moment
braindump:

xxl code snippets are evaluated in, and represented by, whats called a context

to see the contents of your current context, enter . at the xxl prompt

contexts are initially created by mkworkspace(), which creates a context based
on rootctx(), which comes from rootctx.h; this is where the standard verbs are
defined

a context is its own type (x), a list, aka vector of value pointers (VPs),
composed of [scope,code]. a context should never be more than 2 elements.

scope is a regular dictionary. things that the code refers to are looked up in here.
these are your variables from the coders perspective.

code is a regular list, though it should be tagged 'lambda, and look like this:
'lambda#[code,2]

where 2 is the arity (if it uses x/y) or 1 for x

when you type something in at the terminal, it just looks it up in your current
context.

if it's not found there, and the current context has a member named 'parent,
it will try to look it up in there. this is recursive.

by default when you evaluate something using "load" its in your current context,
so things set in that source file will appear in your current context. likewise,
anything you have set in your workspace is visible to that file.

later i will add "import", which cannot make changes to your workspace.

when your source code is parsed, things that look like expressions (ie between
{}, (), or []) are grouped into a sublist and tagged with their punctuation
type.

things inside { .. } are code (obviously) and called lambdas. when lambdas are
encountered while evaluating the parse tree, a new context is created.

its parent, simply a key set in the context dictionary, is set to the current
context that was running the parse tree where this was found.

when you save it, or evaluate/apply it, it will run inside this subcontext.

anything assigned with is or as inside a subcontext (lambda/anonymous
function) will affect only itself, and contexts created from it.

however, sometimes you want to modify something in a parent scope. remember
that in XXL you end up having many many scopes because we use that as a
method of flow control.

here's an example:
```0 as 'ctr; 'upd is {ctr+1 as 'ctr}; []upd```

in there, you really want the parent's ctr value to be updated.

so, to handle that, you can use `ctr+1 as '.ctr`. the leading period
means "somewhere up the tree".

it will use get() to find the matching value (same function that handles
identifiers in things you type in at the REPL), and if found, then
replace its contents using xreplace().

the exact implications of this are currently unknown.

erlang would handle this case with a process who handles the state and
accepts updates via messages. i'm not quite there yet, and i think its
convenient if this is transparent to the user, or at least not require a
significant change in program flow.